test/e2e: move retryutil to e2eutils
This also updates DeploymentReplicaCheck slightly to fail on
errors that are not "NotFound" errors
