+++ test/e2e/e2eutil/retry_util.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package retryutil
+package e2eutil
 
 import (
 	"fmt"
